refactor: Move computeFormatedPercentage and averageSpeedKmH in helpers
