+++ src/lib/helpers.js
@@ -51,3 +51,29 @@ export const formatDate = ({ f, lang, date }) => {
   }
   return f(date, 'HH:mm')
 }
+
+/**
+ * Compute the percentage and returns it as a formatted string
+ * @param {number} value
+ * @param {number} total
+ * @returns {string}
+ */
+export const computeFormatedPercentage = (value, total) => {
+  if (total === 0) return '0%'
+
+  const [int, dec] = ((value * 100) / total).toFixed(2).split('.')
+  if (dec === '00') return `${int}%`
+
+  return `${int}.${dec}%`
+}
+
+/**
+ * Get the average speed in km/h from an array of m/s values
+ * @param {Array} speeds - The speed values, in m/s
+ * @returns {number} The average speed, given in km/h
+ */
+export const averageSpeedKmH = speeds => {
+  const avgSpeed = speeds.reduce((a, b) => a + b, 0) / speeds.length
+  // The speed is given in m/s
+  return avgSpeed * 3.6
+}
