+++ src/lib/trips.spec.js
@@ -5,12 +5,7 @@ import {
   modeProps
 } from 'test/mockTrip'
 
-import {
-  getSectionsFromTrip,
-  getPurpose,
-  getFeatureMode,
-  computeFormatedPercentage
-} from 'src/lib/trips'
+import { getSectionsFromTrip, getPurpose, getFeatureMode } from 'src/lib/trips'
 
 const mockedFeatures = () => [
   mockFeature('featureId01'),
@@ -151,21 +146,4 @@ describe('getFeatureMode', () => {
 
     expect(result).toBe('UNKNOWN')
   })
-})
-
-describe('computeFormatedPercentage', () => {
-  it.each`
-    value | total  | result
-    ${50} | ${100} | ${'50%'}
-    ${50} | ${90}  | ${'55.56%'}
-    ${50} | ${80}  | ${'62.50%'}
-    ${50} | ${0}   | ${'0%'}
-    ${0}  | ${0}   | ${'0%'}
-    ${0}  | ${100} | ${'0%'}
-  `(
-    `should return $result with ($value, $total) params`,
-    ({ value, total, result }) => {
-      expect(computeFormatedPercentage(value, total)).toBe(result)
-    }
-  )
 })
